pyproject.toml
[project]
name = "paperweave"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel>=6.30.0",
    "pandas>=2.3.1",
    "neo4j>=5.20.0",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.31.0",
    "lxml>=5.3.0",
    "schedule>=1.2.0",
]
